requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
docopt==0.6.2
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gTTS==2.5.4
h11==0.16.0
idna==3.11
Jinja2==3.1.6
Js2Py==0.74
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==26.0
pipwin==0.5.2
propcache==0.4.1
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyjsparser==2.7.1
PyJWT==2.11.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-multipart==0.0.22
regex==2026.1.15
requests==2.32.5
scipy==1.17.0
webrtcvad==2.0.10
setuptools==82.0.0
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.2
twilio==9.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
yarl==1.22.0
